[#12][fe][schedule]refacotr: 코드 파일 분리
